requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.41
alembic==1.15.2
psycopg2-binary==2.9.10
asyncpg==0.30.0
google-api-python-client==2.169.0
google-auth-oauthlib==1.2.2
google-auth==2.40.2
cryptography==45.0.3
pydantic==2.11.4
pydantic-settings==2.9.1
typing-extensions>=4.12.2
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.5.0
pydantic[email]
